Put stubs into their own files
